:recycle: 仿照sklearn范式重构TollesLawsonCompensator
:recycle: 拆分TollesLawson预测模型和补偿模型
:heavy_plus_sign: 添加typing-extensions以兼容Python 3.11以前的版本中使用Self返回类型
