+++ pyproject.toml
@@ -12,7 +12,8 @@ dependencies = [
     "scipy>=1.11.1", 
     "pandas>=2.0.3", 
     "loguru>=0.7.0", 
-    "scikit-learn>=1.3.0"
+    "scikit-learn>=1.3.0",
+    "typing-extensions"
 ]
 
 # Author Info
